stop bundles timer
why hasn't she stopped this timer yet? She stopped the fireflies one and those were added more recently. Is there an issue with finding an end date? I'd understand that but when you're running a website with all these timers on it it's best to be up-to-date for the people in the back who haven't seen the updates yet.